+++ CHANGES.md
@@ -1,8 +1,9 @@
 # Release Notes
 
-## 0,5,2 (TDB)
+## 0.5.2 (TDB)
 
 * add `tilejson` URL links for `defaults layers` defined in mosaic's metadata in `/mosaic/register` and `/mosaic/{mosaic_id}/info` response
+* support multiple `layers` in `/mosaic/WMTSCapabilities.xml` endpoint
 
 ## 0.5.1 (2023-08-03)